feat(data table manager): update data table import
