+++ packages/components/data-table-manager/data-table-manager-context/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "dist/commercetools-uikit-data-table-manager-data-table-manager-context.cjs.js",
+  "module": "dist/commercetools-uikit-data-table-manager-data-table-manager-context.esm.js"
+}
