--- packages/components/data-table-manager/src/context/index.ts
@@ -1,8 +1,0 @@
-export {
-  DataTableManagerProvider,
-  DataTableManagerContext,
-} from './data-table-manager-context';
-export type {
-  TDataColumns,
-  TDataTableManagerContext,
-} from './data-table-manager-context';
