+++ packages/components/data-table-manager/src/data-table-manager-context/index.ts
@@ -1,0 +1,5 @@
+export { default } from './data-table-manager-context';
+export type {
+  TDataTableManagerContext,
+  TDataColumns,
+} from './data-table-manager-context';
