+++ packages/components/data-table-manager/src/index.ts
@@ -1,8 +1,5 @@
-export {
-  default,
-  DataTableManagerContext,
-  DataTableManagerProvider,
-} from './data-table-manager';
+export { default } from './data-table-manager';
+export { default as DataTableManagerContext } from './data-table-manager-context/data-table-manager-context';
 export { UPDATE_ACTIONS } from './constants';
 export { default as version } from './version';
 export * from './export-types';
